requirement.txt
click==6.7
Django==2.0.2
django-contrib-comments==1.8.0
django-pagedown==1.0.4
django-pygments==0.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mistune==0.8.3
Pillow==5.0.0
Pygments==2.2.0
pytz==2017.3
Werkzeug==0.14.1
